organising/adding mvvm on sign in view
